Change the auth layout structure
